Section_A/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
apache-beam[gcp]
apache-airflow-providers-google
pytest
